# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GeneSilico EHR API",
        "summary": "Ontada (iKnowMed) FHIR + Stedi eligibility for Texas Oncology.",
...
        "contact": {
            "name": "GeneSilico",
            "email": "developers@genesilico.ai"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8090",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/admin/cancer-centers": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "List Centers",
                "operationId": "list_centers_api_admin_cancer_centers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CancerCenterList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Create Center",
                "operationId": "create_center_api_admin_cancer_centers_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CancerCenterIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CancerCenterOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/hospitals": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "List Hospitals",
                "operationId": "list_hospitals_api_admin_hospitals_get",
                "parameters": [
                    {
                        "name": "cancer_center_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Cancer Center Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HospitalList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Create Hospital",
                "operationId": "create_hospital_api_admin_hospitals_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HospitalIn"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HospitalOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CancerCenterIn": {
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "title": "Name"
                    },
                    "city": {
                        "type": "string",
                        "title": "City",
                        "default": ""
                    },
                    "state": {
                        "type": "string",
                        "title": "State",
                        "default": ""
                    },
                    "organization_id": {
                        "type": "string",
                        "title": "Organization Id",
                        "default": ""
                    },
                    "ehr_vendor": {
                        "type": "string",
                        "title": "Ehr Vendor",
                        "default": "none"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "default": "active"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "CancerCenterIn"
            },
            "CancerCenterList": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/CancerCenterOut"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "title": "CancerCenterList"
            },
            "CancerCenterOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "organizationId": {
                        "type": "string",
                        "title": "Organizationid"
                    },
                    "ehrVendor": {
                        "type": "string",
                        "title": "Ehrvendor"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "createdAt": {
                        "type": "string",
                        "title": "Createdat"
                    },
                    "hospitalIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Hospitalids"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "city",
                    "state",
                    "organizationId",
                    "ehrVendor",
                    "status",
                    "createdAt",
                    "hospitalIds"
                ],
                "title": "CancerCenterOut"
            },
...
            "ErrorBody": {
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code",
                        "description": "Machine-readable: unconfigured, upstream_error, ontada_not_connected, form_invalid, not_implemented \u2026"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Human-readable. Upstream messages are kept verbatim."
                    },
                    "detail": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "ErrorBody"
            },
            "ErrorEnvelope": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorEnvelope",
                "description": "Every non-2xx response from this API has exactly this shape."
            },
...
            "HospitalIn": {
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "title": "Name"
                    },
                    "cancer_center_id": {
                        "type": "string",
                        "title": "Cancer Center Id"
                    },
                    "city": {
                        "type": "string",
                        "title": "City",
                        "default": ""
                    },
                    "state": {
                        "type": "string",
                        "title": "State",
                        "default": ""
                    },
                    "npi": {
                        "type": "string",
                        "title": "Npi",
                        "default": ""
                    },
                    "beds": {
                        "type": "integer",
                        "title": "Beds",
                        "default": 0
                    },
                    "type": {
                        "type": "string",
                        "title": "Type",
                        "default": "hospital"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "default": "active"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cancer_center_id"
                ],
                "title": "HospitalIn"
            },
            "HospitalList": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/HospitalOut"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "title": "HospitalList"
            },
            "HospitalOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cancerCenterId": {
                        "type": "string",
                        "title": "Cancercenterid"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "npi": {
                        "type": "string",
                        "title": "Npi"
                    },
                    "beds": {
                        "type": "integer",
                        "title": "Beds"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "cancerCenterId",
                    "city",
                    "state",
                    "npi",
                    "beds",
                    "type",
                    "status"
                ],
                "title": "HospitalOut"
            },
...
        }
    },
...
}